client/requirements.txt
python-socketio[client]
pyautogui
pytest
pytest-asyncio
